+++ src/demo/custom-editors.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { CustomEditorProps, CustomFilterEditorProps, EditorWrapperProps } from "../library";
 import { Pokemon, query } from "./db";
+import { createMultiSelect } from "./multi-select";
 
 interface DBPokemonType {
   name: string;
@@ -8,6 +9,87 @@ interface DBPokemonType {
 
 interface TypeEditorState {
   options: DBPokemonType[];
+}
+
+export class CustomMultiTypeFilterEditor extends React.Component<CustomFilterEditorProps, TypeEditorState> {
+  state = {
+    options: [] as DBPokemonType[]
+  }
+
+  el: HTMLSelectElement | null = null;
+  prevEl: HTMLSelectElement | null = null;
+
+  constructor(props: CustomFilterEditorProps) {
+    super(props);
+
+    this.changeRef = this.changeRef.bind(this);
+  }
+
+  componentDidMount() {
+    // initialize dropdown
+    let types: DBPokemonType[] = query('SELECT * FROM pokemon_types;');
+    this.setState({ options: types });
+  }
+
+  componentWillUnmount() {
+    this.clearMultiselect();
+  }
+
+  clearMultiselect() {
+    // @ts-ignore
+    if (!!this.el && typeof this.el.destory_multiselect === 'function') {
+      // @ts-ignore
+      this.el.destory_multiselect();
+    }
+
+    // @ts-ignore
+    let div: HTMLDivElement | null | undefined = this.el?.multiselectdiv;
+    if (div) {
+      div.remove();
+    }
+
+    this.props.inputRef.current = null;
+  }
+
+  changeRef(el: HTMLSelectElement | null) {
+    console.log({
+      prev: this.prevEl,
+      curr: this.el,
+      next: el
+    });
+    if (this.el !== el) {
+      // unload prev
+      this.clearMultiselect();
+      this.prevEl = this.el;
+    }
+
+    this.el = el;
+    if (!!el) {
+      // @ts-ignore
+      if (typeof el.loadOptions !== 'function') {
+        createMultiSelect(el);
+
+        // @ts-ignore
+        this.props.inputRef.current = el.multiselectdiv;
+      }
+    }
+  }
+
+  render() {
+    const { value, setValue, editorOptions } = this.props;
+    const { options } = this.state; //
+
+    const allOptions = (editorOptions?.additionalOptions ?? []).concat(options);
+
+    return <select value={value ?? []} onChange={(e) => {
+      let selOptions = Array.from(e.target.selectedOptions).map(o => o.value);
+      setValue(selOptions);
+    }} multiple ref={this.changeRef}>
+      {allOptions.map(t => {
+        return <option key={t.name} value={t.name}>{t.name}</option>
+      })}
+    </select>
+  }
 }
 
 export class CustomTypeSelectEditor extends React.Component<CustomFilterEditorProps, TypeEditorState> {