+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - Pass row edit data to custom cell editors - allows altering other editors based on changes to other cells.
 - `setValues` function to custom editors to allow updating multiple columns at once (ex. side effect changes).
+- Filter editor support for custom single field multi value entries (multi-select dropdowns)
 
 ### Fixed
 - Scroll to top on page change with fixed headers